pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "primordial-soup"
version = "0.1.0"
description = "Governance simulation study for long-term organizational value creation"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "mrg32k3a>=2.0",
    "pyyaml>=6.0",
    "pyarrow>=14.0",
    "matplotlib>=3.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[tool.setuptools.packages.find]
where = ["src"]

# --- Ruff ---

[tool.ruff]
line-length = 99
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # bugbear
    "SIM",  # simplify
    "TCH",  # type-checking imports
]

[tool.ruff.lint.isort]
known-first-party = ["primordial_soup"]

# --- Mypy ---

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# --- Pytest ---

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
